refactor(visualization): move advisor prompt logic into advisor
